Fix home page episode time display and implement user preference-based time formatting
- Move date/time formatting filters from app_tags.py to user_tags.py where they belong
- Add new user_event_time filter for formatting event times according to user preferences
- Update home_grid.html to use user_event_time instead of display_time property
- Update calendar templates to use user_event_time filter
- Add user object to all relevant view contexts (import_data, integrations, export_data, about, history_modal, calendar, list_detail)
- Fix template loading issues by ensuring user object is available in all contexts
- All episode time displays now respect user's time format preferences
