+++ src/templates/events/components/calendar_grid.html
@@ -28,7 +28,7 @@
                   {% with color=release.item.media_type|media_color %}
                     {% icon release.item.media_type is_active=False extra_classes=color|add:" w-4 h-4 flex-shrink-0" %}
                   {% endwith %}
-                  <span class="truncate" title="{{ release }} {{ release.display_time }}">{{ release }} {{ release.display_time }}</span>
+                  <span class="truncate" title="{{ release }} {{ release|user_event_time:user }}">{{ release }} {{ release|user_event_time:user }}</span>
                 </div>
               </a>
             {% endfor %}
