+++ src/templates/events/components/calendar_list.html
@@ -23,7 +23,7 @@
                 <div class="flex items-center">
                   <span class="text-xs text-gray-400 me-2">{{ release.item.media_type|media_type_readable }}</span>
                   {% if release.readable_content_number %}
-                    <span class="text-xs text-indigo-400">{{ release.readable_content_number }} {{ release.display_time }}</span>
+                    <span class="text-xs text-indigo-400">{{ release.readable_content_number }} {{ release|user_event_time:user }}</span>
                   {% endif %}
                 </div>
               </div>
